requirements.txt
# Core dependencies
numpy>=1.20.0,<2.0.0
pandas>=1.3.0,<2.1.0

# Quantitative libraries
# Qlib (Microsoft) - install from GitHub
git+https://github.com/microsoft/qlib.git
akshare>=1.12.0
lightgbm>=3.3.0

# NLP and ML
transformers>=4.30.0
torch>=2.0.0

# Testing
pytest>=7.0.0
pyyaml>=6.0
